+++ src/decoder.rs
@@ -1048,12 +1048,12 @@ impl AvifDecoder {
             ];
             let rgb_stride = width as u32 * 4;
             match planes.chroma_sampling() {
-                ChromaSampling::Cs420 => crate::yuv_bilinear_fix::yuv420_bilinear_complete(
+                ChromaSampling::Cs420 => yuv::yuv420_to_rgba_bilinear(
                     &planar,
                     rgb::bytemuck::cast_slice_mut(out.as_mut_slice()),
                     rgb_stride,
-                    4,
-                    |p, o, s| yuv::yuv420_to_rgba_bilinear(p, o, s, yuv_range, matrix),
+                    yuv_range,
+                    matrix,
                 ),
                 ChromaSampling::Cs422 => yuv::yuv422_to_rgba_bilinear(
                     &planar,
@@ -1085,12 +1085,12 @@ impl AvifDecoder {
             let mut out = vec![Rgb { r: 0u8, g: 0, b: 0 }; pixel_count];
             let rgb_stride = width as u32 * 3;
             match planes.chroma_sampling() {
-                ChromaSampling::Cs420 => crate::yuv_bilinear_fix::yuv420_bilinear_complete(
+                ChromaSampling::Cs420 => yuv::yuv420_to_rgb_bilinear(
                     &planar,
                     rgb::bytemuck::cast_slice_mut(out.as_mut_slice()),
                     rgb_stride,
-                    3,
-                    |p, o, s| yuv::yuv420_to_rgb_bilinear(p, o, s, yuv_range, matrix),
+                    yuv_range,
+                    matrix,
                 ),
                 ChromaSampling::Cs422 => yuv::yuv422_to_rgb_bilinear(
                     &planar,
@@ -1264,26 +1264,26 @@ impl AvifDecoder {
             ];
             let rgb_stride = width as u32 * 4;
             match (planes.chroma_sampling(), bit_depth) {
-                (ChromaSampling::Cs420, 10) => crate::yuv_bilinear_fix::yuv420_bilinear_complete(
+                (ChromaSampling::Cs420, 10) => yuv::i010_to_rgba10_bilinear(
                     &planar,
                     rgb::bytemuck::cast_slice_mut(out.as_mut_slice()),
                     rgb_stride,
-                    4,
-                    |p, o, s| yuv::i010_to_rgba10_bilinear(p, o, s, yuv_range, matrix),
+                    yuv_range,
+                    matrix,
                 ),
-                (ChromaSampling::Cs420, 12) => crate::yuv_bilinear_fix::yuv420_bilinear_complete(
+                (ChromaSampling::Cs420, 12) => yuv::i012_to_rgba12_bilinear(
                     &planar,
                     rgb::bytemuck::cast_slice_mut(out.as_mut_slice()),
                     rgb_stride,
-                    4,
-                    |p, o, s| yuv::i012_to_rgba12_bilinear(p, o, s, yuv_range, matrix),
+                    yuv_range,
+                    matrix,
                 ),
-                (ChromaSampling::Cs420, _) => crate::yuv_bilinear_fix::yuv420_bilinear_complete(
+                (ChromaSampling::Cs420, _) => yuv::i016_to_rgba16_bilinear(
                     &planar,
                     rgb::bytemuck::cast_slice_mut(out.as_mut_slice()),
                     rgb_stride,
-                    4,
-                    |p, o, s| yuv::i016_to_rgba16_bilinear(p, o, s, yuv_range, matrix),
+                    yuv_range,
+                    matrix,
                 ),
                 (ChromaSampling::Cs422, 10) => yuv::i210_to_rgba10(
                     &planar,
@@ -1350,26 +1350,26 @@ impl AvifDecoder {
             ];
             let rgb_stride = width as u32 * 3;
             match (planes.chroma_sampling(), bit_depth) {
-                (ChromaSampling::Cs420, 10) => crate::yuv_bilinear_fix::yuv420_bilinear_complete(
+                (ChromaSampling::Cs420, 10) => yuv::i010_to_rgb10_bilinear(
                     &planar,
                     rgb::bytemuck::cast_slice_mut(out.as_mut_slice()),
                     rgb_stride,
-                    3,
-                    |p, o, s| yuv::i010_to_rgb10_bilinear(p, o, s, yuv_range, matrix),
+                    yuv_range,
+                    matrix,
                 ),
-                (ChromaSampling::Cs420, 12) => crate::yuv_bilinear_fix::yuv420_bilinear_complete(
+                (ChromaSampling::Cs420, 12) => yuv::i012_to_rgb12_bilinear(
                     &planar,
                     rgb::bytemuck::cast_slice_mut(out.as_mut_slice()),
                     rgb_stride,
-                    3,
-                    |p, o, s| yuv::i012_to_rgb12_bilinear(p, o, s, yuv_range, matrix),
+                    yuv_range,
+                    matrix,
                 ),
-                (ChromaSampling::Cs420, _) => crate::yuv_bilinear_fix::yuv420_bilinear_complete(
+                (ChromaSampling::Cs420, _) => yuv::i016_to_rgb16_bilinear(
                     &planar,
                     rgb::bytemuck::cast_slice_mut(out.as_mut_slice()),
                     rgb_stride,
-                    3,
-                    |p, o, s| yuv::i016_to_rgb16_bilinear(p, o, s, yuv_range, matrix),
+                    yuv_range,
+                    matrix,
                 ),
                 (ChromaSampling::Cs422, 10) => yuv::i210_to_rgb10(
                     &planar,